Expanded evalTaylor (Taylor and TaylorN) and also ^ (TaylorN)
evalTaylor can now evaluate a Taylor on a TaylorN, i.e., an implementation
of the composition for Taylor. For TaylorN it evaluates the polynomial
on a vector of coefficients, or also on a vector of Taylor variables.
The latter is implemented on `TaylorSeries.jl`, since it requires that
both utils_Taylor1.jl and utils_TaylorN.jl are loaded.