TaylorRec{T,N}: Implementation as a recursive dense representation with tests
The idea is to generate a polynomial in N vars as a polynomial
of one variable, with coeffs that are polynomials of N-1 variables.

For the time being, it is incorporated in TaylorSeries. It seems to be much
slower, probably due to the bad memory management that the recursion imposses.