WIP: improve evalTaylor for TaylorN
`evalTaylor` (as it is now) yields *not so accurate* results due to accumulation of
round-off errors when used with `TaylorN`. Below, I implement a different
algorithm, which essentially consists on applying Horner's rule for each variable.

So far, this has only be tested in 2-variable expansions; currently it returns
a `TaylorN` object

Functions are not exported