Rename evalTaylor to evaluate and use Horner's method.
WIP: improve evalTaylor for TaylorN

`evalTaylor` (as it is now) yields *not so accurate* results due to accumulation of
round-off errors when used with `TaylorN`. Below, I implement a different
algorithm, which essentially consists on applying Horner's rule for each variable.

So far, this has only be tested in 2-variable expansions; currently it returns
a `TaylorN` object

Functions are not exported

Add @compat for compatibility with v0.3 and correct test

Small change to rerun travis, since tests pass locally

Rename evalTaylor to evaluate, and define horner with different methods

This incorporates some suggestions of David

Use new version of evaluate, delete older and adapt tests