Replace Mathematica .nb by a .m script, and adapt fateman.jl
This saves LOT OF lines, and seems to be the correct way of comparing
the time of the calculation. The time returned is the minimum time
of performing 5 times the calculation.

To be fair, fateman.jl also computes 5 times and returns the minimum
one. A fair comparison of what Mathematica does, one has to compare with
fateman2 or fateman4 output.