Drop support of Julia v0.4
Remove Compat.jl from REQUIRES, and usage of `@compat`, since it is no longer
needed. (It was used to avoid warnings from Julia0.4). Update .travis.yml

Two warnings related to the previous use of `[x=>i for (i,x) in enumerate(v)]`
in `hash_tables.jl` are now solved through the usage of
`Dict(x=>i for (i,x) in enumerate(v))` (Julia 0.5). This is the
actual change that makes dropping support of Julia v0.4 necessary.