Use `fixorder` again, to avoid the if's used in `getindex`
The use of those if's is at the heart of the problem reported in [#102].
The idea was to avoid fixorder by setting the coefficients that are
not defined to zero. This involves if's that killed performance.

Also, fixorder was corrected to return a copy of the coefficients to
avoid side effects.

`max_order` is not anylonger needed, so it is deleted.