Smaller fixes
After creating something like a 5th PR on the example documentation, I think I finally have gotten the hang of how the documenter functions work. This PR clears up some inconsistency (2nd and 3rd order expansion), and fixes the fact that the displayBigO function was set to false, before settinging it to false in the second example. This PR also removed the redundant definition of x halfway through the second example