Add compat for Aqua and stdlibs
TaylorSeries.jl did not specify a compat for Aqua. Aqua 0.8 introduces
some breaking changes, such as renaming test_piracy to test_piracies,
leading to broken test if one has a recent version of Aqua in their
environment.

I added a compat for Aqua. New versions of Aqua also check for compats
for standard libraries (which will be a requirement to publish the
package), so I added those as well.